+++ src/repr/adjacency_matrix/mod.rs
@@ -131,6 +131,7 @@ use crate::{
     Indegree,
     IsComplete,
     IsSimple,
+    IsTournament,
     Order,
     OutNeighbors,
     OutNeighborsWeighted,
@@ -642,6 +643,20 @@ impl IsSimple for AdjacencyMatrix {
     // representation.
     fn is_simple(&self) -> bool {
         self.vertices().all(|u| !self.has_arc(u, u))
+    }
+}
+
+impl IsTournament for AdjacencyMatrix {
+    fn is_tournament(&self) -> bool {
+        let order = self.order();
+
+        if self.size() != order * (order - 1) / 2 {
+            return false;
+        }
+
+        (0..order).all(|u| {
+            (u + 1..order).all(|v| self.has_arc(u, v) ^ self.has_arc(v, u))
+        })
     }
 }