+++ src/repr/edge_list/mod.rs
@@ -123,6 +123,7 @@ use {
         Indegree,
         IsComplete,
         IsSimple,
+        IsTournament,
         Order,
         OutNeighbors,
         OutNeighborsWeighted,
@@ -562,6 +563,20 @@ impl IsSimple for EdgeList {
     // arcs.
     fn is_simple(&self) -> bool {
         self.vertices().all(|u| !self.has_arc(u, u))
+    }
+}
+
+impl IsTournament for EdgeList {
+    fn is_tournament(&self) -> bool {
+        let order = self.order();
+
+        if self.size() != order * (order - 1) / 2 {
+            return false;
+        }
+
+        (0..order).all(|u| {
+            (u + 1..order).all(|v| self.has_arc(u, v) ^ self.has_arc(v, u))
+        })
     }
 }